+++ modules/dnn/CMakeLists.txt
@@ -235,17 +235,17 @@ ocv_list_filterout_ex(dnn_plugin_srcs
 )
 
 if(HAVE_CUDA)
-  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/conv2d_naive.cu")
-  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/relu_naive.cu")
-  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/add_naive.cu")
-  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/pool_naive.cu")
-  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/fc_naive.cu")
-  list(APPEND dnn_int_hdrs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/conv_naive.hpp")
-  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/conv2d_naive.cu" PROPERTIES LANGUAGE CUDA)
-  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/relu_naive.cu" PROPERTIES LANGUAGE CUDA)
-  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/add_naive.cu" PROPERTIES LANGUAGE CUDA)
-  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/pool_naive.cu" PROPERTIES LANGUAGE CUDA)
-  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/fc_naive.cu" PROPERTIES LANGUAGE CUDA)
+  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/conv_cudnn.cu")
+  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/activation_cudnn.cu")
+  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/binary_ops_cudnn.cu")
+  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/pooling_cudnn.cu")
+  list(APPEND dnn_srcs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/matmul_cudnn.cu")
+  list(APPEND dnn_int_hdrs "${CMAKE_CURRENT_LIST_DIR}/src/cuda/layer_cudnn.hpp")
+  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/conv_cudnn.cu" PROPERTIES LANGUAGE CUDA)
+  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/activation_cudnn.cu" PROPERTIES LANGUAGE CUDA)
+  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/binary_ops_cudnn.cu" PROPERTIES LANGUAGE CUDA)
+  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/pooling_cudnn.cu" PROPERTIES LANGUAGE CUDA)
+  set_source_files_properties("${CMAKE_CURRENT_LIST_DIR}/src/cuda/matmul_cudnn.cu" PROPERTIES LANGUAGE CUDA)
 endif()
 
 ocv_option(OPENCV_DNN_OPENVINO "Build with OpenVINO support (2021.4+)" (TARGET ocv.3rdparty.openvino))
