+++ modules/dnn/src/cuda/activation_cudnn.cu
@@ -1,6 +1,6 @@
 #include <cuda_runtime.h>
 #include <float.h>
-#include "conv_naive.hpp"
+#include "layer_cudnn.hpp"
 #include <cudnn.h>
 #include <cuda_fp16.h>
 #include <cuda_bf16.h>
