+++ modules/dnn/src/net_impl2.cpp
@@ -6,7 +6,7 @@
 
 #include "net_impl.hpp"
 #ifdef HAVE_CUDA
-#include "cuda/conv_naive.hpp"
+#include "cuda/layer_cudnn.hpp"
 #include <cuda_runtime.h>
 #include <cudnn.h>
 #endif
